refactor: split Navbar component
